+++ CHANGELOG.md
@@ -91,6 +91,8 @@ and this project adheres to
 
 - Add custom metric to track the number of finalised runs.
   [#1790](https://github.com/OpenFn/lightning/issues/1790)
+- Allow for automatic resubmission of failed usage tracking report submissions.
+  [1789](https://github.com/OpenFn/lightning/issues/1789)
 
 ### Changed